requirements.txt
Pillow
numpy
sympy
pycryptodome
pyzipper
qrcode
pyzbar
rich